+++ packages/f-services/src/index.js
@@ -6,12 +6,14 @@
 
 import axiosServices from './axios';
 import globalisationServices from './globalisation';
+import utilities from './utilities';
 import validations from './validations';
 import windowServices from './window';
 
 export {
     axiosServices,
     globalisationServices,
+    utilities,
     validations,
     windowServices
 };